nivel2/nivel2_app/requirements.txt
Flask==1.1.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.0
mysql-connector
